Changed correlation_functions directory location
