+++ halotools/empirical_models/ia_models/__init__.py
@@ -3,3 +3,4 @@ from .ia_strength_models import *
 from .nfw_phase_space import *
 from .occupation_models import *
 from .watson_distribution import DimrothWatson
+from .correlation_functions import *
