+++ halotools/empirical_models/ia_models/correlation_functions/ed_3d_one_two_halo_decomp.py
@@ -1,6 +1,6 @@
 r"""
-Module containing the `~halotools.mock_observables.alignments.ee_3d` function used to
-calculate the ellipticity-ellipticity (EE) correlation functon
+Module containing the `~halotools.mock_observables.alignments.ed_3d_one_two_halo_decomp` function used to
+calculate the 1-halo and 2-halo contributions to the ellipticity-direction (ED) correlation functon.
 """
 
 from __future__ import absolute_import, division, print_function, unicode_literals
@@ -9,29 +9,28 @@ import numpy as np
 from math import pi
 
 from .alignment_helpers import process_3d_alignment_args
-from ..mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
+from ....mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
-from ..mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
-from ..mock_observables.pair_counters import marked_npairs_3d, npairs_3d, marked_npairs_3d
-from .ee_3d import _ee_3d_process_args
+from ....mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
+from ....mock_observables.pair_counters import positional_marked_npairs_3d, npairs_3d, marked_npairs_3d
+from .ed_3d import _ed_3d_process_args
 
-__all__ = ['ee_3d_one_two_halo_decomp']
+__all__ = ['ed_3d_one_two_halo_decomp']
 __author__ = ['Duncan Campbell']
 
 
 np.seterr(divide='ignore', invalid='ignore')  # ignore divide by zero in e.g. DD/RR
 
 
-def ee_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
-                              sample2, orientations2, sample2_host_halo_id,
-                              rbins, 
+def ed_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
+                              sample2, sample2_host_halo_id, rbins,
                               weights1=None, weights2=None,
                               mask1=None, mask2=None,
                               period=None, num_threads=1,
                               approx_cell1_size=None, approx_cell2_size=None):
     r"""
-    Calculate the one and two halo componenents of the 3-D ellipticity-ellipticity
-    correlation function (EE), :math:`\eta_{\rm 1h}(r)`, and :math:`\eta_{\rm 2h}(r)`.
+    Calculate the one and two halo componenents of the 3-D ellipticity-direction
+    correlation function (ED), :math:`\omega_{\rm 1h}(r)`, and :math:`\omega_{\rm 2h}(r)`.
 
     Parameters
     ----------
@@ -43,31 +42,35 @@ def ee_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
 
     orientations1 : array_like
-        Npts1 x 2 numpy array containing projected orientation vectors for each point in ``sample1``.
+        Npts1 x 3 numpy array containing orientation vectors for each point in ``sample1``.
         these will be normalized if not already.
 
     sample1_host_halo_id : array_like
-        *len(sample1)* integer array of host halo ids.
+        Npts1 length integer array of host halo ids.
 
     sample2 : array_like, optional
         Npts2 x 3 array containing 3-D positions of points.
 
-    orientations2 : array_like
-        Npts1 x 3 numpy array containing orientation vectors for each point in ``sample2``.
-        these will be normalized if not already.
-
     sample2_host_halo_id : array_like
-        *len(sample2)* integer array of host halo ids.
+        Npts2 length integer array of host halo ids.
 
     rbins : array_like
         array of boundaries defining the radial bins in which pairs are counted.
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
 
-     weights1 : array_like, optional
-        Npts1 array of weghts.  If this parameter is not specified, it is set to numpy.ones(Npts1).
+    weights1 : array_like, optional
+        Npts1 array of weights.  If this parameter is not specified, it is set to numpy.ones(Npts1).
 
     weights2 : array_like, optional
-        Npts2 array of weghts.  If this parameter is not specified, it is set to numpy.ones(Npts2).
+        Npts2 array of weights.  If this parameter is not specified, it is set to numpy.ones(Npts2).
+
+    mask1 : array_like, optional
+        Npts1 boolean array indicating which galaxies in `sample1` contributes to the ED correlation function.
+        The default is np.array([True]*Npts1).
+
+    mask2 : array_lile, optional
+        Npts2 boolean array indicating which galaxies in `sample2` contributes to the ED correlation function.
+        The default is np.array([True]*Npts2).
 
     period : array_like, optional
         Length-3 sequence defining the periodic boundary conditions
@@ -100,20 +103,21 @@ def ee_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
 
     Returns
     -------
-    correlation_function : numpy.array
-        *len(rbins)-1* length array containing the correlation function :math:`\eta(r)`
-        computed in each of the bins defined by input ``rbins``.
+    correlation_functions : numpy.array
+        Two *len(rbins)-1* length array containing the correlation function :math:`\omega_{1\rm h}(r)`
+        and :math:`\omega_{2\rm h}(r)` computed in each of the bins defined by input ``rbins``.
 
     Notes
     -----
-    The ellipticity-ellipticity correlation function is defined as:
+    The ellipticity-direction correlation function is defined as:
 
     .. math::
-        \eta = \frac{\sum_{i \neq j}w_iw_j|\hat{e}_i \cdot \hat{e}_j|^2}{\sum_{i \neq j} w_i w_j} - \frac{1}{3}
+        \omega = \frac{\sum_{i \neq j}w_iw_j|\hat{e}_i \cdot \hat{r}_{ij}|^2}{\sum_{i \neq j} w_iw_j} - \frac{1}{3}
 
-    where e.g. :math:`\hat{e}_i` is the orientation of the :math:`i`-th galaxy.
-    :math:`w_i` and :math:`w_j` are the weights associated with the :math:`i`-th
-    and :math:`j`-th galaxy. The weights default to 1 if not set.
+    where e.g. :math:`\hat{e}_i` is the orientation of the :math:`i`-th galaxy, and
+    :math:`\hat{r}_{ij}` is the normalized vector in the direction of the :math:`j`-th galaxy
+    from the :math:`i`-th galaxy.  :math:`w_i` and :math:`w_j` are the weights associated with
+    the :math:`i`-th and :math:`j`-th galaxy. The weights default to 1 if not set.
 
     Example
     --------
@@ -145,24 +149,24 @@ def ee_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
 
     >>> halo_ids = np.random.randint(1, 10, Npts)
 
-    We can the calculate the auto-EE correlation between these points:
+    We can the calculate the auto-ED correlation between these points:
 
     >>> rbins = np.logspace(-1,1,10)
-    >>> result_1h, result_2h = ee_3d_one_two_halo_decomp(sample1, random_orientations, halo_ids, sample1, random_orientations, halo_ids, rbins, period=Lbox)
+    >>> result = ed_3d_one_two_halo_decomp(sample1, random_orientations,  halo_ids, sample1, halo_ids, rbins, period=Lbox)
 
     """
 
     # process arguments
     alignment_args = (sample1, orientations1, None, weights1,
-                      sample2, orientations2, None, weights2,
+                      sample2, None, None, weights2,
                       None, None, None, None)
     dum = 0.0  # dummy variable to store arguments not needed for this function
     sample1, orientations1, dum, weights1,\
-    sample2, orientations2, dum, weights2,\
+    sample2, dum, dum, weights2,\
     dum, dum, dum, dum = process_3d_alignment_args(*alignment_args)
 
     function_args = (sample1, rbins, sample2, period, num_threads)
-    sample1, rbins, sample2, period, num_threads, PBCs = _ee_3d_process_args(*function_args)
+    sample1, rbins, sample2, period, num_threads, PBCs = _ed_3d_process_args(*function_args)
 
     # How many points are there (for normalization purposes)?
     N1 = len(sample1)
@@ -202,23 +206,22 @@ def ee_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
     marks1[:,2] = orientations1[:,1]
     marks1[:,3] = orientations1[:,2]
     marks1[:,4] = sample1_host_halo_id
-    marks2 = np.zeros((N2, 5))
+    marks2 = np.zeros((N2, 2))
     marks2[:,0] = weights2
-    marks2[:,1] = orientations2[:,0]
-    marks2[:,2] = orientations2[:,1]
-    marks2[:,3] = orientations2[:,2]
-    marks2[:,4] = sample2_host_halo_id
+    marks2[:,1] = sample2_host_halo_id
 
-    marked_counts_1h = marked_npairs_3d(sample1[mask1], sample2[mask2], rbins,
-                                        period=period, weights1=marks1[mask1], weights2=marks2[mask2],
-                                        weight_func_id=16, num_threads=num_threads,
-                                        approx_cell1_size=approx_cell1_size, approx_cell2_size=approx_cell2_size)
+    marked_counts_1h, counts = positional_marked_npairs_3d(sample1[mask1], sample2[mask2], rbins,
+                                period=period, weights1=marks1[mask1], weights2=marks2[mask2],
+                                weight_func_id=7, num_threads=num_threads,
+                                approx_cell1_size=approx_cell1_size,
+                                approx_cell2_size=approx_cell2_size)
     marked_counts_1h = np.diff(marked_counts_1h)
 
-    marked_counts_2h = marked_npairs_3d(sample1[mask1], sample2[mask2], rbins,
-                                        period=period, weights1=marks1[mask1], weights2=marks2[mask2],
-                                        weight_func_id=17, num_threads=num_threads,
-                                        approx_cell1_size=approx_cell1_size, approx_cell2_size=approx_cell2_size)
+    marked_counts_2h, counts = positional_marked_npairs_3d(sample1[mask1], sample2[mask2], rbins,
+                                period=period, weights1=marks1[mask1], weights2=marks2[mask2],
+                                weight_func_id=8, num_threads=num_threads,
+                                approx_cell1_size=approx_cell1_size,
+                                approx_cell2_size=approx_cell2_size)
     marked_counts_2h = np.diff(marked_counts_2h)
 
     # if no weights, use fast un-weighted pair counter
@@ -257,10 +260,8 @@ def ee_3d_one_two_halo_decomp(sample1, orientations1, sample1_host_halo_id,
                        approx_cell2_size=approx_cell2_size)
     counts_2h = np.diff(counts_2h)
 
-    #counts = counts_1h + counts_2h
-
-    result_1h =  marked_counts_1h/counts - 1.0/3.0*(counts_1h/counts)
-    result_2h =  marked_counts_2h/counts - 1.0/3.0*(counts_2h/counts)
+    result_1h = marked_counts_1h/counts - 1.0/3.0*(counts_1h/counts)
+    result_2h = marked_counts_2h/counts - 1.0/3.0*(counts_2h/counts)
 
     return result_1h, result_2h
