+++ halotools/empirical_models/ia_models/correlation_functions/ee_projected.py
@@ -9,10 +9,10 @@ import numpy as np
 from math import pi
 
 from .alignment_helpers import process_projected_alignment_args
-from ..mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
+from ....mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
-from ..mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
-from ..mock_observables.pair_counters import marked_npairs_xy_z, npairs_xy_z, marked_npairs_xy_z
+from ....mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
+from ....mock_observables.pair_counters import marked_npairs_xy_z, npairs_xy_z, marked_npairs_xy_z
 
 __all__ = ['ee_projected']
 __author__ = ['Duncan Campbell']
