+++ halotools/empirical_models/ia_models/correlation_functions/gi_plus_3d.py
@@ -9,10 +9,10 @@ import numpy as np
 from math import pi, gamma
 
 from .alignment_helpers import process_3d_alignment_args
-from ..mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
+from ....mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
-from ..mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
-from ..mock_observables.pair_counters import positional_marked_npairs_3d, marked_npairs_3d
+from ....mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
+from ....mock_observables.pair_counters import positional_marked_npairs_3d, marked_npairs_3d
 
 __all__ = ['gi_plus_3d']
 __author__ = ['Duncan Campbell']
