+++ halotools/empirical_models/ia_models/correlation_functions/ii_minus_3d.py
@@ -1,41 +1,33 @@
 r"""
-Module containing the `~halotools.mock_observables.alignments.ii_minus_projected` function used to
-calculate the projected intrinsic ellipticity-ellipticity (II) correlation
+Module containing the `~halotools.mock_observables.alignments.ii_minus_3d` function used to
+calculate the intrinsic ellipticity-ellipticity (II) correlation
 """
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
-from math import pi
+from math import pi, gamma
 
-from .alignment_helpers import process_projected_alignment_args
-from ..mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
+from .alignment_helpers import process_3d_alignment_args
+from ....mock_observables.mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
-from ..mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
-from ..mock_observables.pair_counters import positional_marked_npairs_xy_z, marked_npairs_xy_z
+from ....mock_observables.pair_counters.mesh_helpers import _enforce_maximum_search_length
+from ....mock_observables.pair_counters import positional_marked_npairs_3d, marked_npairs_3d
 
-__all__ = ['ii_minus_projected']
+__all__ = ['ii_plus_3d']
 __author__ = ['Duncan Campbell']
 
 
 np.seterr(divide='ignore', invalid='ignore')  # ignore divide by zero in e.g. DD/RR
 
 
-def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientations2, ellipticities2,
-            rp_bins, pi_max, randoms1=None, randoms2=None, weights1=None, weights2=None,
+def ii_minus_3d(sample1, orientations1, ellipticities1, sample2, orientations2, ellipticities2,
+            rbins, randoms1=None, randoms2=None, weights1=None, weights2=None,
             ran_weights1=None, ran_weights2=None, estimator='Natural',
             period=None, num_threads=1, approx_cell1_size=None, approx_cell2_size=None):
     r"""
-    Calculate the projected intrinsic ellipticity-ellipticity correlation function (II),
-    :math:`w_{--}(r_p)`, where :math:`r_p` is the separation perpendicular to the line-of-sight (LOS)
-    between two galaxies.  See the 'Notes' section for details of this calculation.
-
-    The first two dimensions define the plane for perpendicular distances.  The third
-    dimension is used for parallel distances, i.e. x,y positions are on the plane of the
-    sky, and z is the redshift coordinate. This is the 'distant observer' approximation.
-
-    Note in particular that the `~halotools.mock_observables.alignments.ii_minus_projected` function does not
-    accept angular coordinates for the input ``sample1`` or ``sample2``.
+    Calculate the intrinsic ellipticity-ellipticity correlation function (II),
+    :math:`\xi_{--}(r)`.  See the 'Notes' section for details of this calculation.
 
     Parameters
     ----------
@@ -48,7 +40,7 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
 
     orientations1 : array_like
-        Npts1 x 2 numpy array containing projected orientation vectors for each point in ``sample1``.
+        Npts1 x 3 numpy array containing projected orientation vectors for each point in ``sample1``.
         these will be normalized if not already.
 
     ellipticities1: array_like
@@ -58,20 +50,16 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
         Npts2 x 3 array containing 3-D positions of points with associated
         orientations and ellipticities.
 
-    orientations12 : array_like
-        Npts1 x 2 numpy array containing projected orientation vectors for each point in ``sample2``.
+    orientations2 : array_like
+        Npts1 x 3 numpy array containing projected orientation vectors for each point in ``sample2``.
         these will be normalized if not already.
 
     ellipticities2: array_like
         Npts1 x 1 numpy array containing ellipticities for each point in ``sample2``.
 
-    rp_bins : array_like
-        array of boundaries defining the radial bins perpendicular to the LOS in which
+    rbins : array_like
+        array of boundaries defining the radial bins in which
         pairs are counted.
-        Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
-
-    pi_max : float
-        maximum LOS distance defining the projection integral length-scale in the z-dimension.
         Length units are comoving and assumed to be in Mpc/h, here and throughout Halotools.
 
     randoms1 : array_like, optional
@@ -133,22 +121,16 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
     Returns
     -------
     correlation_function : numpy.array
-        *len(rp_bins)-1* length array containing the correlation function :math:`w_{--}(r_p)`
-        computed in each of the bins defined by input ``rp_bins``.
+        *len(rbins)-1* length array containing the correlation function :math:`w_{g+}(r)`
+        computed in each of the bins defined by input ``rbins``.
 
     Notes
     -----
 
-    The projected II-correlation function is calculated as:
+    The II-correlation function is calculated as:
 
     .. math::
-        w_{--}(r_p) = 2 \int_0^{\pi_{\rm max}} \xi_{--}(r_p, \pi) \mathrm{d}\pi
-
-
-    If the Natural estimator is indicated, the projected II-correlation function is calculated as:
-
-    .. math::
-        \xi_{--}(r_p, \pi) = \frac{S_{-}S_{-}}{R_sR_s}
+        \xi_{--}(r) = \frac{S_{-}S_{-}}{R_sR_s}
 
     where
 
@@ -162,11 +144,11 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
         e_{-}(j|i) = e_j\sin(2\phi)
 
     where :math:`e_j` is the ellipticity of the :math:`j`-th galaxy.  :math:`\phi` is the angle between the
-    orientation vector, :math:`\vec{o}_j`, and the projected direction between the :math:`j`-th
-    and :math:`i`-th galaxy, :math:`\vec{r}_{p i,j}`.
+    orientation vector, :math:`\vec{o}_j`, and the direction between the :math:`j`-th
+    and :math:`i`-th galaxy, :math:`\vec{r}_{i,j}`.
 
     .. math::
-        \cos(\phi) = \vec{o}_j \cdot \vec{r}_{p i,j}
+        \cos(\phi) = \vec{o}_j \cdot \vec{r}_{i,j}
 
     :math:`R_sR_s` are random pair counts,
 
@@ -194,14 +176,13 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
 
     We then create a set of random orientation vectors and ellipticities for each point
 
-    >>> random_orientations = np.random.random((Npts,2))
+    >>> random_orientations = np.random.random((Npts,3))
     >>> random_ellipticities = np.random.random(Npts)
 
     We can the calculate the projected auto-GI correlation between these points:
 
-    >>> rp_bins = np.logspace(-1,1,10)
-    >>> pi_max = 0.25
-    >>> w = ii_minus_projected(sample1, random_orientations, random_ellipticities, sample1, random_orientations, random_ellipticities, rp_bins, pi_max, period=Lbox)
+    >>> rbins = np.logspace(-1,1,10)
+    >>> w = ii_minus_3d(sample1, random_orientations, random_ellipticities, sample1, random_orientations, random_ellipticities, rbins, period=Lbox)
 
     """
 
@@ -211,12 +192,12 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
                       randoms1, ran_weights1, randoms2, ran_weights2)
     sample1, orientations1, ellipticities1, weights1, sample2,\
         orientations2, ellipticities2, weights2, randoms1, ran_weights1,\
-        randoms2, ran_weights2 = process_projected_alignment_args(*alignment_args)
+        randoms2, ran_weights2 = process_3d_alignment_args(*alignment_args)
 
-    function_args = (sample1, rp_bins, pi_max, sample2, randoms1, randoms2,
+    function_args = (sample1, rbins, sample2, randoms1, randoms2,
         period, num_threads, approx_cell1_size, approx_cell2_size)
-    sample1, rp_bins, pi_bins, sample2, randoms1, randoms2,\
-        period, num_threads, PBCs, no_randoms = _ii_minus_projected_process_args(*function_args)
+    sample1, rbins, sample2, randoms1, randoms2,\
+        period, num_threads, PBCs, no_randoms = _ii_minus_3d_process_args(*function_args)
 
     # How many points are there (for normalization purposes)?
     N1 = len(sample1)
@@ -230,35 +211,36 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
 
     #define merk vectors to use in pair counting
     # sample 1
-    marks1 = np.ones((N1, 3))
+    marks1 = np.ones((N1, 4))
     marks1[:, 0] = ellipticities1 * weights1
     marks1[:, 1] = orientations1[:, 0]
     marks1[:, 2] = orientations1[:, 1]
+    marks1[:, 3] = orientations1[:, 2]
     # sample 2
-    marks2 = np.ones((N2, 3))
+    marks2 = np.ones((N2, 4))
     marks2[:, 0] = ellipticities2 * weights2
-    marks2[:, 1] = orientations1[:, 0]
-    marks2[:, 2] = orientations1[:, 1]
+    marks2[:, 1] = orientations2[:, 0]
+    marks2[:, 2] = orientations2[:, 1]
+    marks2[:, 3] = orientations2[:, 2]
     # randoms 1
-    ran_marks1 = np.ones((NR1, 3))
+    ran_marks1 = np.ones((NR1, 4))
     ran_marks1[:, 0] = ran_weights1
     ran_marks1[:, 1] = 0  # dummy
     ran_marks1[:, 2] = 0  # dummy
+    ran_marks1[:, 3] = 0  # dummy
     # randoms 2
-    ran_marks2 = np.ones((NR2, 3))
+    ran_marks2 = np.ones((NR2, 4))
     ran_marks2[:, 0] = ran_weights2
     ran_marks2[:, 1] = 0  # dummy
     ran_marks2[:, 2] = 0  # dummy
-
-    # define pi bins
-    pi_bins = np.array([0.0, pi_max])
+    ran_marks2[:, 3] = 0  # dummy
 
     do_SS, do_RR = II_estimator_requirements(estimator)
 
     # count marked pairs
     if do_SS:
         SS = marked_pair_counts(sample1, sample2,  marks1,  marks2,
-                                rp_bins, pi_bins, period, num_threads,
+                                rbins, period, num_threads,
                                 approx_cell1_size, approx_cell2_size)
     else:
         SS = None
@@ -266,19 +248,19 @@ def ii_minus_projected(sample1, orientations1, ellipticities1, sample2, orientat
     # count random pairs
     if do_RR:
         RR = random_counts(randoms1, randoms2, ran_weights1, ran_weights2,
-                           rp_bins, pi_bins, N1, N2, no_randoms, period, PBCs,
+                           rbins, N1, N2, no_randoms, period, PBCs,
                            num_threads, approx_cell1_size, approx_cell2_size)
     else:
         RR = None
 
     result = II_estimator(SS, RR, N1, N2, NR1, NR2, estimator)
 
-    return result*2.0*pi_max  # factor of 2pi_max accounts for integration
+    return result  # factor of 2pi_max accounts for integration
 
 
 def II_estimator(SS, RR, N1, N2, NR1, NR2, estimator='Natural'):
     r"""
-    apply the supplied II estimator to calculate the correlation function.
+    apply the supplied GI estimator to calculate the correlation function.
     """
     if estimator == 'Natural':
         factor = (NR1*NR2)/(N1*N2)
@@ -290,7 +272,7 @@ def II_estimator(SS, RR, N1, N2, NR1, NR2, estimator='Natural'):
 
 def II_estimator_requirements(estimator):
     r"""
-    Return the requirments for the supplied II estimator.
+    Return the requirments for the supplied GI estimator.
     """
 
     do_RR = False
@@ -305,24 +287,23 @@ def II_estimator_requirements(estimator):
         raise ValueError(msg)
 
 
-def marked_pair_counts(sample1, sample2, weights1, weights2, rp_bins, pi_bins, period,
+def marked_pair_counts(sample1, sample2, weights1, weights2, rbins, period,
         num_threads, approx_cell1_size, approx_cell2_size):
     r"""
     Count marked pairs.
     """
 
     weight_func_id = 6
-    SS = positional_marked_npairs_xy_z(sample1, sample2, rp_bins, pi_bins, period=period,
+    SS = positional_marked_npairs_3d(sample1, sample2, rbins, period=period,
         weights1=weights1, weights2=weights2, weight_func_id=weight_func_id,
         num_threads=num_threads, approx_cell1_size=approx_cell1_size,
         approx_cell2_size=approx_cell1_size)[0]
-    SS = np.diff(np.diff(SS, axis=0), axis=1)
-    SS = SS.flatten()
+    SS = np.diff(SS, axis=0)
 
     return SS
 
 
-def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rp_bins, pi_bins,
+def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rbins,
                   N1, N2, no_randoms, period,
                   PBCs, num_threads, approx_cell1_size, approx_cell2_size):
     r"""
@@ -330,13 +311,12 @@ def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rp_bins, pi_bi
     """
 
     if no_randoms is False:
-        RR = marked_npairs_xy_z(randoms1, randoms2, rp_bins, pi_bins,
+        RR = marked_npairs_3d(randoms1, randoms2, rbins,
                 period=period, num_threads=num_threads, weight_func_id=1,
                 weights1=ran_weights1, weights2=ran_weights2,
                 approx_cell1_size=approx_cell1_size,
                 approx_cell2_size=approx_cell2_size)
-        RR = np.diff(np.diff(RR, axis=0), axis=1)
-        RR = RR.flatten()
+        RR = np.diff(RR, axis=0)
 
         return RR
     else:
@@ -346,8 +326,8 @@ def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rp_bins, pi_bi
         NR2 = N2
 
         # do volume calculations
-        v = cylinder_volume(rp_bins, 2.0*pi_bins)
-        dv = np.diff(np.diff(v, axis=0), axis=1)
+        v = nball_volume(rbins)
+        dv = np.diff(v, axis=0)
         global_volume = period.prod()
 
         # calculate the random-random pairs.
@@ -357,18 +337,19 @@ def random_counts(randoms1, randoms2, ran_weights1, ran_weights2, rp_bins, pi_bi
         return RR.flatten()
 
 
-def cylinder_volume(R, h):
-    r"""
-    Calculate the volume of a cylinder(s), used for the analytical randoms.
+def nball_volume(R, k=3):
     """
-    return pi*np.outer(R**2.0, h)
+    Calculate the volume of a n-shpere.
+    This is used for the analytical randoms.
+    """
+    return (np.pi**(k/2.0)/gamma(k/2.0+1.0))*R**k
 
 
-def _ii_minus_projected_process_args(sample1, rp_bins, pi_max, sample2, randoms1, randoms2,
+def _ii_minus_3d_process_args(sample1, rbins, sample2, randoms1, randoms2,
         period, num_threads, approx_cell1_size, approx_cell2_size):
     r"""
     Private method to do bounds-checking on the arguments passed to
-    `~halotools.mock_observables.alignments.ii_minus_projected`.
+    `~halotools.mock_observables.alignments.alignments.ii_plus_projected`.
     """
     sample1 = enforce_sample_has_correct_shape(sample1)
 
@@ -395,15 +376,12 @@ def _ii_minus_projected_process_args(sample1, rp_bins, pi_max, sample2, randoms1
     else:
         no_randoms = False
 
-    pi_max = float(pi_max)
-    pi_bins = np.array([0.0, pi_max])
-
-    rp_bins = get_separation_bins_array(rp_bins)
-    rp_max = np.amax(rp_bins)
+    rbins = get_separation_bins_array(rbins)
+    rmax = np.amax(rbins)
 
     period, PBCs = get_period(period)
 
-    _enforce_maximum_search_length([rp_max, rp_max, pi_max], period)
+    _enforce_maximum_search_length([rmax, rmax, rmax], period)
 
     if (randoms1 is None) & (PBCs is False):
         msg = "If no PBCs are specified, both randoms must be provided.\n"
@@ -411,4 +389,5 @@ def _ii_minus_projected_process_args(sample1, rp_bins, pi_max, sample2, randoms1
 
     num_threads = get_num_threads(num_threads)
 
-    return sample1, rp_bins, pi_bins, sample2, randoms1, randoms2, period, num_threads, PBCs, no_randoms
+    return sample1, rbins, sample2, randoms1, randoms2, period, num_threads, PBCs, no_randoms
+
