+++ halotools/empirical_models/ia_models/correlation_functions/tests/test_ed_projected.py
