+++ halotools/empirical_models/ia_models/correlation_functions/tests/test_gi_plus_projected.py
