+++ halotools/empirical_models/ia_models/correlation_functions/tests/test_ii_plus_projected.py
